Replace WorkActivitiesByDate with ClientTasksList for client tasks view
